vendor: pick up latest cilium
This reduces the number of dependencies and pulls in the uint64 ->
uint32 change in flow protobuf definition (see [0]).

[0]: https://github.com/cilium/cilium/pull/11161

+++ cmd/observe/observe_filter_test.go
@@ -162,8 +162,8 @@ func TestIdentity(t *testing.T) {
 	cmd := newObserveCmd(f)
 	require.NoError(t, cmd.Flags().Parse([]string{"--identity", "1", "--identity", "2"}))
 	assert.Equal(t, []*pb.FlowFilter{
-		{SourceIdentity: []uint64{1, 2}},
-		{DestinationIdentity: []uint64{1, 2}},
+		{SourceIdentity: []uint32{1, 2}},
+		{DestinationIdentity: []uint32{1, 2}},
 	}, f.whitelist.flowFilters())
 	assert.Nil(t, f.blacklist)
 }
@@ -172,7 +172,7 @@ func TestFromIdentity(t *testing.T) {
 	f := newObserveFilter()
 	cmd := newObserveCmd(f)
 	require.NoError(t, cmd.Flags().Parse([]string{"--from-identity", "1", "--from-identity", "2"}))
-	assert.Equal(t, []*pb.FlowFilter{{SourceIdentity: []uint64{1, 2}}}, f.whitelist.flowFilters())
+	assert.Equal(t, []*pb.FlowFilter{{SourceIdentity: []uint32{1, 2}}}, f.whitelist.flowFilters())
 	assert.Nil(t, f.blacklist)
 }
 
@@ -180,7 +180,7 @@ func TestToIdentity(t *testing.T) {
 	f := newObserveFilter()
 	cmd := newObserveCmd(f)
 	require.NoError(t, cmd.Flags().Parse([]string{"--to-identity", "1", "--to-identity", "2"}))
-	assert.Equal(t, []*pb.FlowFilter{{DestinationIdentity: []uint64{1, 2}}}, f.whitelist.flowFilters())
+	assert.Equal(t, []*pb.FlowFilter{{DestinationIdentity: []uint32{1, 2}}}, f.whitelist.flowFilters())
 	assert.Nil(t, f.blacklist)
 }
